pyproject.toml
[project]
name = "snake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
]
